[feature|optimize|fix] Support sharing and copying images from reading page; support recording playback progress; optimize player code; fix some issues

+++ app/src/main/java/com/skyd/anivu/model/repository/download/DownloadRepository.kt
@@ -39,8 +39,10 @@ class DownloadRepository @Inject constructor(
         link: String,
     ): Flow<Unit> {
         return flow {
-            downloadManager.getTorrentFilesByLink(link = link).forEach {
-                File(it.path).deleteRecursively()
+            if (downloadManager.getDownloadState(link)?.downloadComplete() != true) {
+                downloadManager.getTorrentFilesByLink(link).forEach {
+                    File(it.path).deleteRecursively()
+                }
             }
             val requestUuid = downloadManager.getDownloadInfo(link)?.downloadRequestId
             if (!requestUuid.isNullOrBlank()) {
